Get existing or create channel id for given topic and viewpoint
This implements the core functionality to create a channel id or get an
existing one to match users with opposing viewpoints. If there is no
existing channel with the same topic and opposing viewpoint then it adds
a new channel to a queue. Once there's a request for the same topic and
opposing viewpoint the original channel is pulled from the queue (in a
FIFO manner).