Refactor channel retriever to not use a static data store
Using a static data store like this can cause unexpected behavior when
it's used in the same JVM. Particularly it was causing tests to fail.
However, it's also bad design. Intead of having it be static we now
instantiate it on application start up and pass it into the retriever.
This way there's only one instance but it provides the benefits
mentioned above.