Allow application to access dynamodb
The application didn't have credentials to access dynamodb. This change
passes the two necessary keys starting from github secrets and ending up
as environment variables passed into the docker container. As part of
this process we add configuration to the application to accept these
values as envars. We also configure ansible to run the container with
these vars.

- Refactor twilio and aws secret configuration to the same build
function.
- Add more checks that variables were assigned before trying to use them
in the build scripts.